requirements.txt
streamlit
python-dotenv
langchain
langchain-openai
langchain-community
faiss-cpu
unstructured
